+++ generators/control/templates/control.plain.js
@@ -1,0 +1,13 @@
+<%- banner %>
+define( [], function() {
+  'use strict';
+
+   return function create() {
+      var element = document.createElement( 'DIV' );
+
+      /* :) */
+
+      return element;
+   };
+
+} );